+++ projects/demo-maps/src/app/route-components/route-example-custom-layers/customRenderers/threejs_renderer.ts
@@ -4,7 +4,8 @@ import LayerRenderer from 'ol/renderer/Layer';
 import VectorLayer from 'ol/layer/Vector';
 import { Polygon as olPolygonGeometry } from 'ol/geom';
 import {
-  WebGLRenderer, PerspectiveCamera, Scene, Mesh, Renderer, DirectionalLight, MeshPhongMaterial, Vector3, Shape, ExtrudeGeometry, Color
+  WebGLRenderer, PerspectiveCamera, Scene, Mesh, Renderer,
+  DirectionalLight, MeshPhongMaterial, Vector3, Shape, ExtrudeGeometry, Color, BufferGeometry
 } from 'three';
 import { heightAboveWidth } from '@dlr-eoc/map-three';
 
@@ -128,7 +129,7 @@ export class ThreeJsRenderer extends LayerRenderer<VectorLayer> {
     return meshes;
   }
 
-  private fGeom2tGeom(fGeom: olPolygonGeometry, height: number) {
+  private fGeom2tGeom(fGeom: olPolygonGeometry, height: number): BufferGeometry {
     const coords = fGeom.getCoordinates()[0];
     const shape = new Shape();
     shape.moveTo(coords[0][0], coords[0][1]);