+++ projects/services-ogc/src/lib/services-ogc.module.ts
@@ -1,12 +1,14 @@
 import { NgModule } from '@angular/core';
-import { OwcJsonService } from './owc/owc-json.service';
 import { WmtsClientService } from './wmts/wmtsclient.service';
+import { WmsService } from './wms/wmsclient.service';
 import { WpsClient } from './wps/wpsclient';
+import { OwcJsonService } from './owc/owc-json.service';
+
 
 @NgModule({
   declarations: [],
   imports: [],
   exports: [],
-  providers: [OwcJsonService, WmtsClientService, WpsClient]
+  providers: [WmtsClientService, WmsService, OwcJsonService, WpsClient]
 })
 export class ServicesOgcModule { }