+++ projects/services-ogc/src/public-api.ts
@@ -8,4 +8,4 @@ export * from './lib/owc/types/owc-json';
 export * from './lib/owc/types/eoc-owc-json';
 export * from './lib/wps/wpsclient';
 export * from './lib/wmts/wmtsclient.service';
-export * from './lib/wms/wms.service';
+export * from './lib/wms/wmsclient.service';