+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+# [7.3.0](https://github.com/dlr-eoc/ukis-frontend-libraries/tree/v7.3.0) (2021-04-20) (Layer, Utils Features and bug fixes maps)
 ### Bug Fixes
 * **@dlr-eoc/layer-control:**
   - adjust misleading title on hover state [issue 53](https://github.com/dlr-eoc/ukis-frontend-libraries/issues/53).
@@ -9,6 +10,10 @@
   - removeLayerByKey() not working on layergroups (recursive).
   - addLayer(), addLayers() and setLayers() don't set the filtertype on layergroup children (recursive).
   - `updateLayerParamsWith` now also works with GeojsonLayers: [issue 51](https://github.com/dlr-eoc/ukis-frontend-libraries/issues/51)
+* **@dlr-eoc/map-three:**
+  - regarding three-version 125: `Geometry` is now deprecated, replaced with `BufferGeometry`
+
+
 
 ### Features
 * **@dlr-eoc/layer-control:**:
@@ -21,6 +26,7 @@
   - now allows for angular-components in popups. In any Ukis-Layer, just add a `dynamicPopup` field to the `popup` property.
     If the popup-component requires any `@Input`'s, provide these with `getAttributes(args: IPopupArgs)`. 
     Addresses [Issue #14](https://github.com/dlr-eoc/ukis-frontend-libraries/issues/14)
+  - added basic functionality for WMS-layers to map-ol. As for now, `VectorLayers` of type `wfs` do not have any dedicated own methods or properties. All wfs-specific methods, like CQL-filters, are simply red from the layer's `url` property.
 * **@dlr-eoc/utils-maps:**:
   - created new custom renderer `InterpolationRenderer`.
     - Takes a vector-source (or a clustered vector-source) and does inverse-distance-interpolation on every pixel between the data points.
@@ -34,7 +40,9 @@
     - Original implementation remains as a wrapper for backwards-compatibility.
     - Added method `describeProcess`
     - Added ability to overwrite data-decoding with a user-provided method, for cases where a unconventional mime-type or encoding has been used.
-  - Expose WmsService [Pull #70](https://github.com/dlr-eoc/ukis-frontend-libraries/pull/70).
+    - Expose WmsService [Pull #70](https://github.com/dlr-eoc/ukis-frontend-libraries/pull/70).
+* **@dlr-eoc/utils-maps:**
+  - updated all custom renderers to implement the method `renderDeclutter` that is required as of ol6.5.
 
 ### Refactoring
 * **@dlr-eoc/map-ol:**